Stop rejecting valid mathematics: km, non-integer number lines, and a grid window the question can fit in
Three schema limits were rejecting correct questions, and between them they are
most of why the four figure-bearing topics ran at 20-38% acceptance while the
rest ran at 60-100%.

triangleLabeled and quadrilateralLabeled allowed cm, m and mm. Bearings and
navigation questions are set in kilometres in every paper, and eleven drafts
died on the enum.

numberLine required integer bounds. An inequality line from -2.5 to 2.5 is
ordinary, and the template already carried its own `step`, so the ticks were
never tied to whole-number bounds in the first place.

coordinateGrid capped each range at 20 units. The mesh step scales with the
span, so a wider window stays readable; a graph of cost against 30 items, or
distance over 40 km, was being rejected for being the size the question is. Now
40.

None of these were judgement calls about difficulty or style — they were our
own limits refusing mathematics the syllabus sets.

+++ lib/visuals/templates/numberLine.ts
@@ -5,8 +5,11 @@ import { valueStatedInText, type VisualTemplate } from '../types';
 // Number line with tick marks, optional labeled points, and optional
 // solution-set ray/segment (for inequalities).
 export const NumberLineParamsZ = z.object({
-  min: z.number().int().min(-100).max(100),
-  max: z.number().int().min(-100).max(100),
+  // Not integers: an inequality line from -2.5 to 2.5 is ordinary, and the
+  // template already carries its own `step`, so the ticks were never tied to
+  // whole-number bounds.
+  min: z.number().min(-100).max(100),
+  max: z.number().min(-100).max(100),
   step: z.number().positive().max(50).default(1),
   points: z
     .array(
